Refactoring housekeeper
We basically have 3 types of job:
* Jobs that are executed regularly and which callback don't depend on anything
  (records, comments)
* Jobs that are executed regularly, but which the callback depend on a specific interval
  (snooze, notification, audits)
* The backup job (which dumps the db)

In order to make it more pallatable, it's easier to make class for each type of job.
Each job has a `next_run` attribute which is computed everytime it runs, and contains
the date of the next run. The loop to check becomes very simple this way.

This refactor will make it easy to expand, and easier to test.
Additionally we made the Housekeeper not depend on core, but only on the database. This
was done to make it easier to test.