Tweaking the reload system
Let's try to remove as much references to `cluster` in modules.
The cluster should handle the case where it's alone on his own,
instead of having a `if` condition for every call.

Also refactoring `reload_data` into `load_data` and `reload_data`.
`reload_data` is meant to be overwritten, and `load_data` is only
for usage in `reload_data`.
This removes the useless `sync` argument in `reload_data`.
Also made the `sync_neighbors` method for usage in `reload_data`
to be DRY.