Fixing kombu issue when the URL is too long
UnicodeError: encoding with 'idna' codec failed (UnicodeError: label empty or too long)

This happens when the URL passed to kombu is too long.
To solve this, instead of recomputing a connection string to pass to kombu, we're making
a kombu transport/channel that support MongoDB database object as an input.
We can thus reuse the same connections and settings (along with settings unsupported by the native kombu
implementation).