[config] Making a dedicated route for schema
Weirdly enough, settings was having 2 purposes:
* Reading/writing settings at `/etc/snooze/server/<section>.yaml`
* Reading web form schema at `snooze/defaults/web/<endpoint>.yaml`

This was working out because of the previous config API that we were
using, which basically read/write any file on demand. This was also
creating a potential security threat, since the user may craft a
query to read/write an arbitrary file.

With the current config API, this is not possible anymore. Let's
separate the two functionalities and give the schema reading a
dedicated route.