Database call optimization for increments
Increments for statistics and number of hits should be written at
regular intervals instead of for every record processed.
The intermediate value can stay in memory.

Also getting rid of the complex Database() returning a snooze.db.<type>.BackendDB
instance, since it's very counter intuitive, and it's not really possible to have
the parent returning the children in python.