More precise exception catching
No need to throw a warning upon KeyError, but dateutil.parser.ParseError
is interesting to throw.
We let the system error when an OverflowError is raised, because it is
unexpected and should return a 5xx. We should technically catch it better
to return a 400, but that should do it for now. Also, if the user overflow the
maximum date number storable in C, it's technically not a bad request.