Refactoring logging
We should separate logging into 2:
* API logging: generally used for human interactions
* process logging: Used for incoming alerts and their process

There is a far bigger quantity of logs for process than api, so it
makes sense that they should be rotated differently.
Also, the process logs need to include the hash or uid of the record,
and this needs to be in a standardized format.