Better logs for reloading
Most plugin execute actions after reloading, and thus are calling
super().reload_data() before executing their custom code.
We would like to have info logs when a plugin reload is performed, so
we can see the action of a PUT/POST/DELETE on the in-memory plugin
value.
To make it easier, we made a new _post_reload method that can be
overwritten. All plugin that were overwritting reload_data will now
overwrite _post_reload to make it easier (except the rule plugin, which
really overwrite reload_data), and to be able to have logs before and
after each reload operation.