Fixing a bug where a new search would return no result
When searching a new thing, we will be placed in the current page
we're searching for for the new search.
For instance, if we're on page 3, and the new search doesn't have
3 pages of result, then nothing will be displayed, making the user
think there is no result for his search.
Every search should reset the page number, that is the expected user
behavior.