More abstraction
Less code in common/opensearch, moved to apiserver/.
It's too complicated to have model-related functional code in
common/opensearch. It's not flexible, and when one change is done
on apiserver, we need to change common/opensearch code.
It's much simpler to define helpers in common/opensearch, and then
copy-paste queries between apiserver and processor, so the 2 components
don't impact each other.